Kinematics

Julia's main source-to-source reverse mode automatic differentiation package (Zygote.jl) does not allow for mutation of array elements. At the same time, the dual-number based forward-mode automatic differentiation in ForwardDiff.jl is most efficient when used with non-allocating code. As a compromise, the kinematics functions in DiscreteElasticRods.jl come in both non-mutating and non-allocating forms. The latter are distinguished by a trailing !, and usually require a cache argument.

Basic 3-Vector Operations

DiscreteElasticRods.cross3 — Function

    cross3(a,b)

Cross products of the columns of two 3 x N arrays. Output is also 3 x N

DiscreteElasticRods.dot3 — Function

    dot3(a,b)

Dot products of the columns of two 3 x N arrays. Output is 1 x N

DiscreteElasticRods.norm3 — Function

    norm3(a,b)

Norms of the columns of two 3 x N arrays. Output is 1 x N

Low-Level Rod Operations

DiscreteElasticRods.edges — Function

    edges(x)

Edges of a centerline x, stored as a 3 x N array. Output is 3 x N-1

DiscreteElasticRods.tangents — Function

    tangents(x)

Tangents of a centerline x, stored as a 3 x N array. Output is 3 x N-1

DiscreteElasticRods.rotate_orthogonal_unit — Function

    rotate_orthogonal_unit(v, t, θ)

Rotate unit vectors v around axes t1 by angles θ.

v and t are (3 x N), θ is (1 x N).

The columns v[:,i] and t[:,i] must be orthogonal unit vectors. The function uses the Rodriguez rotation formula speciailized to this case.

DiscreteElasticRods.ptransport — Function

    ptransport(v, t1, t2)

Parallel transport of vectors v from tangents t1 to tangents 't2'.

All inputs are (3 x N). Each column of t1 and t2 must have unit norm.

Uses the Rodriguez rotation formula. There is a singularity in this definition when t1 == t2, which is handled by switching to an small-θ expansion (see ptransportinnercoefficient).

High-Level Rod Operations

DiscreteElasticRods.full_kinematics — Function

    full_kinematics(x, d)

Returns the edge lengths, curvatures, and twists of a rod.

Arguments

x: centerline vertices, size (3 x N)

d: directors, size (3 x N-1). Unit length and orthogonal to edges(x)

Returns

l: edge lengths (1 x N-1)

κ: curvature normals (2 x N-2), expressed in the director frame

'τ': scalar twist at each vertex (1 x N-2)

Twists are calculated by parallel transporting the directors across each vertex, and comparing with the original. This avoids the reference twist calculations used in Lestringant, Audoly, and Kochmann (2020)

DiscreteElasticRods.elastic_energy — Function

    elastic_energy(l0,κ0,τ0, l,κ,τ, k,B,β)

Elastic energy, calculated from stretching, bending, and twisting strains.

Arguments

l0, κ0, τ0: reference lengths, curvatures, and twists

l, κ, τ: current lengths, curvatures, and twists

k, B, β: stretching, bending, and twisting stiffness.

Lengths, curvatures, and twists should be (1 x N-1), (2 x N-2), and (1 x N-2), respectively.

k and β may be scalar or similar to l and τ. The bending stiffness B may constant isotropic (scalar), constant anisotropic (2 x 1), or non-constant anisotropic (2 x N-2).

DiscreteElasticRods.gravitational_energy — Function

    gravitational_energy(x, m, g)

Gravitational energy, calculated from centerline.

Arguments

x: centerline vertices, 3 x N

m: edges masses, 1 x N-1. Typically ρ*A*l0

g: magnitude of gravitational acceleration (positive).

Gravity is assumed to act in the negative z direction.