Problem Specification

These functions are used for constructing rods, specifying boundary conditions, and specifying material properties.

DiscreteElasticRods.elliptical_stiffness — Function

elliptical_stiffness(E, ν, a, b)

Return the elastic properties of a rod with elliptical section.

Arguments

E: Elastic Modulus

ν: Poisson Ratio

a, b: Radii of the cross section

DiscreteElasticRods.straight_rod — Function

straight_rod(p1, p2, d, N)

Returns a rod with endpoints p1, p2 and director d with N segments.

If d has a component parallel to p2 - p1, it is projected out.

DiscreteElasticRods.discretize_span — Function

discretize_span(sspan, ns)

Return the vertex and edge locations (sv and se) for sspan divided into ns segments.

Examples

julia> sv, se = discretize_span([0,1], 10)

DiscreteElasticRods.discrete_rod — Function

discrete_rod(xfun, dfun, s, N)

Return a rod with centerline and directors approximating xfun(s) and dfun(s)

Directors and centerline are sampled from the continous curve, and then any tangential component of the directors is removed by projection.

Examples

julia> r = discrete_rod([0,0,0], [1,0,0], [0,0,1], s -> 2, s -> 1, s -> s*(1-s), [0,1])

DiscreteElasticRods.continuous_rod — Function

continuous_rod(x0, t0, d0, κ1, κ2, τ, s)

Returns two functions x and d defining the centerline and directors of a rod.

These are determined by solving the ODE which defines a Bishop frame with curvatures κ1(s), κ2(s), and τ(s). The integration starts from initial conditions x0, t0, and d0, and covers length s = [s1, s2]

Examples

julia> xfun, dfun = continuous_rod([0,0,0], [1,0,0], [0,0,1],
                                      s -> 2, s -> 1, s -> s*(1-s), [0,1])